ACL init, profiles and permissions table

+++ app/Http/Controllers/Admin/PlanController.php
@@ -56,10 +56,19 @@ class PlanController extends Controller
 
     public function delete($url)
     {
-        $plan = $this->repository->where('url',$url)->first();
+        $plan = $this->repository
+                    ->with('details')
+                    ->where('url',$url)
+                    ->first();
 
         if(!$plan)
             return redirect()->back();
+
+        if($plan->details->count() > 0){
+            return redirect()
+                    ->back()
+                    ->with('error','Existem detalhes vinculados ao plano.');
+        }
 
         $plan->delete();
         return redirect()->route('plans.index');
